requirements.txt
Flask
flask_cors
openai
python-dotenv
